centralizing file writing....
