+++ packages/components/nodes/documentloaders/Docx/Docx.ts
@@ -1,9 +1,7 @@
 import { ICommonObject, INode, INodeData, INodeParams } from '../../../src/Interface'
 import { TextSplitter } from 'langchain/text_splitter'
 import { DocxLoader } from 'langchain/document_loaders/fs/docx'
-import path from 'path'
-import { getStoragePath } from '../../../src'
-import fs from 'fs'
+import { getFileFromStorage } from '../../../src'
 
 class Docx_DocumentLoaders implements INode {
     label: string
@@ -66,8 +64,7 @@ class Docx_DocumentLoaders implements INode {
             const chatflowid = options.chatflowid
 
             for (const file of files) {
-                const fileInStorage = path.join(getStoragePath(), chatflowid, file)
-                const fileData = fs.readFileSync(fileInStorage)
+                const fileData = getFileFromStorage(file, chatflowid)
                 const blob = new Blob([fileData])
                 const loader = new DocxLoader(blob)
