+++ packages/components/nodes/tools/OpenAPIToolkit/OpenAPIToolkit.ts
@@ -3,10 +3,7 @@ import { BaseLanguageModel } from '@langchain/core/language_models/base'
 import { OpenApiToolkit } from 'langchain/agents'
 import { JsonSpec, JsonObject } from './core'
 import { ICommonObject, INode, INodeData, INodeParams } from '../../../src/Interface'
-import { getCredentialData, getCredentialParam } from '../../../src'
-import { getStoragePath } from '../../../src'
-import fs from 'fs'
-import path from 'path'
+import { getCredentialData, getCredentialParam, getFileFromStorage } from '../../../src'
 
 class OpenAPIToolkit_Tools implements INode {
     label: string
@@ -63,9 +60,9 @@ class OpenAPIToolkit_Tools implements INode {
         if (yamlFileBase64.startsWith('FILE-STORAGE::')) {
             const file = yamlFileBase64.replace('FILE-STORAGE::', '')
             const chatflowid = options.chatflowid
-            const fileInStorage = path.join(getStoragePath(), chatflowid, file)
-            const fileData = fs.readFileSync(fileInStorage)
+            const fileData = getFileFromStorage(file, chatflowid)
             const utf8String = fileData.toString('utf-8')
+
             data = load(utf8String) as JsonObject
         } else {
             const splitDataURI = yamlFileBase64.split(',')
