+++ packages/components/src/index.ts
@@ -7,3 +7,4 @@ dotenv.config({ path: envPath, override: true })
 export * from './Interface'
 export * from './utils'
 export * from './speechToText'
+export * from './storageUtils'
