+++ packages/components/src/multiModalUtils.ts
@@ -1,7 +1,5 @@
 import { IVisionChatModal, ICommonObject, IFileUpload, IMultiModalOption, INodeData, MessageContentImageUrl } from './Interface'
-import path from 'path'
-import { getStoragePath } from './utils'
-import fs from 'fs'
+import { getFileFromStorage } from './storageUtils'
 
 export const addImagesToMessages = (
     nodeData: INodeData,
@@ -18,10 +16,8 @@ export const addImagesToMessages = (
             for (const upload of imageUploads) {
                 let bf = upload.data
                 if (upload.type == 'stored-file') {
-                    const filePath = path.join(getStoragePath(), options.chatflowid, options.chatId, upload.name)
-
+                    const contents = getFileFromStorage(upload.name, options.chatflowid, options.chatId)
                     // as the image is stored in the server, read the file and convert it to base64
-                    const contents = fs.readFileSync(filePath)
                     bf = 'data:' + upload.mime + ';base64,' + contents.toString('base64')
 
                     imageContent.push({
